MB-8661: implemented ui auth via 10-minute token
UI is now doing POST to /uilogin on login. Which sets 10 minute cookie
with token identifying authorized user.

Auth tokens are stored and managed per-node in new menelaus_ui_auth
service. They are not replicated.

Long-polling pool-details replies also check if token is about to
expire and if so it generates new token that's set into updated cookie.

{auth_cookie, is replaced with normal auth because normal auth is now
cookie-based.

Change-Id: If0b91428a9e6435104c9afdf314c2fd67b4a3ceb